requirements.txt
ib_async>=1.0.0
psycopg[binary]>=3.2.0
python-dotenv>=1.0.0
requests>=2.31.0
